+++ duckscript/src/lib.rs
@@ -124,7 +124,6 @@
 //!
 
 mod expansion;
-mod io;
 pub mod parser;
 mod preprocessor;
 pub mod runner;